+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - **scheduler** add base and max shift parameters for flow-matching samplers  
   - enhance `--optional` flag to pre-install optional packages  
   - add `[lora]` to recognized filename patterns  
+  - add **Apple DepthPro** controlnet processor, thanks @nolbert82  
 - **Fixes**
   - startup error with `--profile` enabled if using `--skip`  
   - restore orig init image for each batch sequence